requirements.txt
astroid==2.0.1
click==6.7
Flask==1.0.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.3
more-itertools==4.2.0
pkg-resources==0.0.0
PyJWT==1.6.4
pylint==2.0.0
pylint-django==0.11.1
pylint-plugin-utils==0.4
pymongo==3.7.1
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
